Minor edits and comments

+++ Source/Heart/Public/Model/HeartGraphNode3D.h
@@ -27,5 +27,5 @@ public:
 
 private:
 	UPROPERTY()
-	float Height;
+	double Height;
 };
